Split tiles deployment and deploy planet (#264)
Part of #231 

Changes:
- Split tile deployment into 6 regions:
  - africa
  - asia
  - europe
  - north-america
  - oceania
  - south-america

Reasoning: serving the entire planet requires more resources. Generating
tiles for the entire planet takes too long, and the resulting tiles
image / deployment becomes too large. Splitting this per region makes
the deployment parallelizable. Every tile server will have 1 vCPU and
256MB memory.
- Create reusable Github Actions workflow to deploy tiles for a certain
region
- Deploy low-zoom tiles (zooms 0-7) with a separate deployment for the
whole planet
- Change nightly data update and deployment to trigger at 22:47 UTC
- This is expected to take the entire night (possibly more) to render
and deploy tiles for the entire planet.
- Change proxy configuration to match incoming tile requests to their
respective tile server using regular expressions matching `z/x/y`
- Remove all map bounds output after import
- Document the deployment

+++ tiles/tiles.sh
@@ -6,6 +6,10 @@
 #  - https://maplibre.org/martin/martin-cp.html
 #  - https://maplibre.org/martin/mbtiles-copy.html
 
+# Input variables:
+#  - BBOX: the bounding box of the region to generate
+#  - TILES: generate only the low/medium zoom tiles if the value is "low-med", and only the high zoom tiles if the value is "high". Otherwise, generate everything
+
 set -e
 
 OUTPUT_DIR="/tiles"
@@ -15,48 +19,52 @@ export MARTIN="martin-cp --config /config/configuration.yml --mbtiles-type flat 
 
 mkdir -p "$OUTPUT_DIR"
 
-echo "Tiles: low-med"
+if [[ "${TILES}" != 'high' ]]; then
+  echo "Tiles: low-med"
 
-rm -f "$OUTPUT_DIR/railway_line_high.mbtiles"
-$MARTIN --min-zoom 0 --max-zoom 7 --source railway_line_high --output-file "$OUTPUT_DIR/railway_line_high.mbtiles"
-mbtiles summary "$OUTPUT_DIR/railway_line_high.mbtiles"
+  rm -f "$OUTPUT_DIR/railway_line_high.mbtiles"
+  $MARTIN --min-zoom 0 --max-zoom 7 --source railway_line_high --output-file "$OUTPUT_DIR/railway_line_high.mbtiles"
+  mbtiles summary "$OUTPUT_DIR/railway_line_high.mbtiles"
 
-rm -f "$OUTPUT_DIR/standard_railway_text_stations_low.mbtiles"
-$MARTIN --min-zoom 0 --max-zoom 6 --source standard_railway_text_stations_low --output-file "$OUTPUT_DIR/standard_railway_text_stations_low.mbtiles"
-mbtiles summary "$OUTPUT_DIR/standard_railway_text_stations_low.mbtiles"
+  rm -f "$OUTPUT_DIR/standard_railway_text_stations_low.mbtiles"
+  $MARTIN --min-zoom 0 --max-zoom 6 --source standard_railway_text_stations_low --output-file "$OUTPUT_DIR/standard_railway_text_stations_low.mbtiles"
+  mbtiles summary "$OUTPUT_DIR/standard_railway_text_stations_low.mbtiles"
 
-rm -f "$OUTPUT_DIR/standard_railway_text_stations_med.mbtiles"
-$MARTIN --min-zoom 7 --max-zoom 7 --source standard_railway_text_stations_med --output-file "$OUTPUT_DIR/standard_railway_text_stations_med.mbtiles"
-mbtiles summary "$OUTPUT_DIR/standard_railway_text_stations_med.mbtiles"
+  rm -f "$OUTPUT_DIR/standard_railway_text_stations_med.mbtiles"
+  $MARTIN --min-zoom 7 --max-zoom 7 --source standard_railway_text_stations_med --output-file "$OUTPUT_DIR/standard_railway_text_stations_med.mbtiles"
+  mbtiles summary "$OUTPUT_DIR/standard_railway_text_stations_med.mbtiles"
+fi
 
-echo "Tiles: high"
+if [[ "${TILES}" != 'low-med' ]]; then
+  echo "Tiles: high"
 
-rm -f /tiles/high.mbtiles
-$MARTIN --min-zoom 8 --max-zoom "$MAX_ZOOM" --source railway_line_high,railway_text_km --output-file /tiles/high.mbtiles
-mbtiles summary /tiles/high.mbtiles
+  rm -f /tiles/high.mbtiles
+  $MARTIN --min-zoom 8 --max-zoom "$MAX_ZOOM" --source railway_line_high,railway_text_km --output-file /tiles/high.mbtiles
+  mbtiles summary /tiles/high.mbtiles
 
-echo "Tiles: standard"
+  echo "Tiles: standard"
 
-rm -f "$OUTPUT_DIR/standard.mbtiles"
-$MARTIN --min-zoom 8 --max-zoom "$MAX_ZOOM" --source standard_railway_turntables,standard_railway_text_stations,standard_railway_grouped_stations,standard_railway_symbols,standard_railway_switch_ref --output-file "$OUTPUT_DIR/standard.mbtiles"
-mbtiles summary "$OUTPUT_DIR/standard.mbtiles"
+  rm -f "$OUTPUT_DIR/standard.mbtiles"
+  $MARTIN --min-zoom 8 --max-zoom "$MAX_ZOOM" --source standard_railway_turntables,standard_railway_text_stations,standard_railway_grouped_stations,standard_railway_symbols,standard_railway_switch_ref --output-file "$OUTPUT_DIR/standard.mbtiles"
+  mbtiles summary "$OUTPUT_DIR/standard.mbtiles"
 
-echo "Tiles: speed"
+  echo "Tiles: speed"
 
-rm -f "$OUTPUT_DIR/speed.mbtiles"
-$MARTIN --min-zoom 8 --max-zoom "$MAX_ZOOM" --source speed_railway_signals --output-file "$OUTPUT_DIR/speed.mbtiles"
-mbtiles summary "$OUTPUT_DIR/speed.mbtiles"
+  rm -f "$OUTPUT_DIR/speed.mbtiles"
+  $MARTIN --min-zoom 8 --max-zoom "$MAX_ZOOM" --source speed_railway_signals --output-file "$OUTPUT_DIR/speed.mbtiles"
+  mbtiles summary "$OUTPUT_DIR/speed.mbtiles"
 
-echo "Tiles: signals"
+  echo "Tiles: signals"
 
-rm -f "$OUTPUT_DIR/signals.mbtiles"
-$MARTIN --min-zoom 8 --max-zoom "$MAX_ZOOM" --source signals_railway_signals,signals_signal_boxes --output-file "$OUTPUT_DIR/signals.mbtiles"
-mbtiles summary "$OUTPUT_DIR/signals.mbtiles"
+  rm -f "$OUTPUT_DIR/signals.mbtiles"
+  $MARTIN --min-zoom 8 --max-zoom "$MAX_ZOOM" --source signals_railway_signals,signals_signal_boxes --output-file "$OUTPUT_DIR/signals.mbtiles"
+  mbtiles summary "$OUTPUT_DIR/signals.mbtiles"
 
-echo "Tiles: electrification"
+  echo "Tiles: electrification"
 
-rm -f "$OUTPUT_DIR/electrification.mbtiles"
-$MARTIN --min-zoom 8 --max-zoom "$MAX_ZOOM" --source electrification_signals --output-file "$OUTPUT_DIR/electrification.mbtiles"
-mbtiles summary "$OUTPUT_DIR/electrification.mbtiles"
+  rm -f "$OUTPUT_DIR/electrification.mbtiles"
+  $MARTIN --min-zoom 8 --max-zoom "$MAX_ZOOM" --source electrification_signals --output-file "$OUTPUT_DIR/electrification.mbtiles"
+  mbtiles summary "$OUTPUT_DIR/electrification.mbtiles"
+fi
 
 echo "Done"
